mtd: vt8500: Repair serial flash driver after pinctrl/gpio changes
With the removal of the GPIO driver, the serial flash driver now fails
to initialize because it can't find a wm,wm8650-gpio node.

Remove this test as we don't need to detect the bootstrap mode - all
Wondermedia SoCs in production have used a serial flash boot rom so
only 'mode 0' is required.

Signed-off-by: Tony Prisk <linux@prisktech.co.nz>
Signed-off-by: Alexey Charkov <alchark@gmail.com>